sfx2: fix SfxViewShell::CheckIPClient_Impl()
Probably the Edit->Plug-in removal removed the wrong branch there since
m_bPlugInsActive was always true before.

This caused embded::UnreachableStateException to be thrown out of a
Timer and that was only handled as a FatalError in desktop::Main,
for example on kde84498-1.html.

(regression from 1da46d960c1cd59ebb8bd1b85fa39008bcaf9d63)

Change-Id: If9e2578f722ea5c5b0821d3d894bb439dbf3c40c