tdf#98238, tdf#99591 reverted old commit
Revert "tdf#53472 Integral Symbols too small for some integrands"

This reverts commit 1c8434e083003c95e83198f485df25b70e6d2d48.

Revert on request from regina, since this commit caused 2 new bugs.

The code of the reverted patch is kept in the file for further
investigation, once it has been confirmed that it works.

The file had 3 commits later, these are not affected by this commit

Change-Id: I5b24a9b88ac909c83115e8c8a430dc768c9effd4
Reviewed-on: https://gerrit.libreoffice.org/24898
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
Reviewed-by: jan iversen <jani@documentfoundation.org>