Related: tdf#99523 two undo actions listed after dragging a slide
to a new place. "Delete Slides" and "Drag and Drop Slides", there
should only be the "Drag and Drop Slides" shown.

There is one more EndUndo than StartUndo. The extra one is the
last one of View::DragFinished. I surmise that this is meant
to match the BegUndo of View::StartDrag which is not called
in this situation.

This doesn't fix tdf#99523, but its the first thing I noticed,
so fix that first.

Change-Id: I3c31a3cfe83f28892112568bacfb25294366d6cf