recognize #REF! particles of invalidated references, tdf#86575 follow-up
... so they result in an invalid reference again producing a #REF! error
instead of a bad string producing #NAME? error.

This way we can handle the invalid #REF!.A1 and similar references that
were wrongly written to ODFF between 2013 and 2016 until 5.1.4

As a benefit, this is now also the case in UI, e.g. when recompiling
changed names with already invalidated references.

Change-Id: I117d709f594b7c37d899528a51220c1855b7817d