tdf#98792: Add the 'Nirmala UI' font as fallback for 'Gautami' for Telugu
Nirmala UI is the new pan-Indic font on Windows 8 and newer, and
apparently not optional. It supports the Devanagari, Bengali,
Gurmukhi, Gujarati, Oriya, Tamil, Telugu, Kannada, Malayalam, and
Sinhala scripts. So we should list it in many more places, I guess,
not just for Telugu.

Note that this commit does not really fix the root issue(s) in
tdf#98792, i.e. why the shaping is wrong when, and/or why the font (or
glyph?)  fallback works differently depending on whether the document
is passed on the command line or opened from the Start Centre. See bug
report for more notes.

Change-Id: Ife361475716123781fb4a4923701890206b28b26