opengl: VertexUtils, deferred flush update, remove unneeded calls
VertexUtils - collection of utils for working with Vertices.

Add deferred flush to some places directly where it is called
indirectly. This is to assure that we don't produce regressions
if we change the behavior in the future.

drawAlphaBitmap is the same as drawBitmap so when drawBitmap is
called just redirect to drawAlphaBitmap

Change-Id: Ibef1ba88865856d92d9e93734cf5d6561af785c0