tdf#100187 fix division by zero in comboFragmentShader
When feather is 0.0 (used when anti-aliasing is disabled) then
we get a "division by zero" situation. As per OpenGL secs. the
shader should not fail in this situation however the result is
undetermined. Most GPUs handled this correctly but on some the
lines didn't draw at all. This should fix this issue.

Change-Id: I56ca2f10c393491807321969c72085ef7690d16a