Resolves: rhbz#1344042 deb#826654 scale gtk3 smooth scaling events...
to give traditional amounts of scroll on a single mouse wheel event

ditch non smooth scroll events now seeing as apparently they are always
available so the other types are irrelevent now

if we get x and y scroll, then like macosx just dispatch x and y scroll events.

Note: there seems to be a bug in the stack below us where the first scroll
event after getting focus is one of a 0 x and y delta.  Because we now check x
and y against 0, we don't launch a scroll event in the case of a 0x0 scroll
which stops us occasionally appearing to go backwards on the first scroll after
getting focus. Which is the same thing I see on e.g. gedit, the first mouse
wheel scroll after getting focus doesn't actually do anything.

Change-Id: Iec8f2e4627cd84e3896270a0847a5c4907fa083f