Resolves: tdf#100313 OutputDevice::isDisposed doesn't work anymore
since

commit 75ea7913ab99d77bbe36a3cf343efc5d896d9367
Author: melikeyurtoglu <aysemelikeyurtoglu@gmail.com>
Date:   Tue May 3 00:34:05 2016 +0300

    tdf#97527 vcl: reference-count Menu

    Change-Id: Ia12434fede69ad247ed67691517437a9ada31acd
    Signed-off-by: melikeyurtoglu <aysemelikeyurtoglu@gmail.com>
    Reviewed-on: https://gerrit.libreoffice.org/24596

which created a base-class VclReferenceBase for OutputDevice and
Menu and so moved isDisposed from OutputDevice to VclReferenceBase,
but *duplicated* the mbDisposed in VclReferenceBase so the
bit set by disposeOnce in OutputDevice was a different bit to the
bit returned by isDisposed from VclReferenceBase.

Which meant that the fix of...

commit d19177614e0b1edda9a05db9eeb08bffc9e9e456
Author: Caolán McNamara <caolanm@redhat.com>
Date:   Fri Nov 27 16:10:10 2015 +0000

    Resolves: rhbz#1283426 using vdevs based on now dead physical devs is unsafe

no longer worked.

Finish moving the mbDisposed to VclReferenceBase to fix this

Change-Id: I9bc1ba3d03f3aae7c3f58eb277176d9521bdb05d