tdf#100275 sw: fix target node of bookmark copy
Replaces the defensive programming band-aid of
aecde7271c4573b82a5f7d46d13a0b2ec1495ef1 with a real fix.

The problem is that lcl_NonCopyCount() has some special case code
to ignore the first node in the target document, which erroneously is
executed for every bookmark, which results in the 2 bookmarks in the
bugdoc being created with nDelCount 1 and 2 so they land on the same
node, which is not allowed for cross-reference marks.

Extract the adjustment into a separate function that is called once.

(regression from 281f893a33cedf002bdba3c823952aa4aabd295f)

Change-Id: Ie14c650f7fdb259c13cb9048226da30971d2ab3c