tdf#98797 sw: try to fix crash in SwMailDispatcherListener_Impl
Commit 7034df6327e9db3b0a7f51c7ab28b49faa1a9647 added a isDisposed()
call in SwMailDispatcherListener_Impl::idle() and i can't see why that
would not be necessary in the other functions that access
m_pSendMailDialog if it is necessary in idle() - it's not obvious *why*
it would be necessary in idle() in the first place but obviously i don't
understand when SwSendMailDialog::dispose() is invoked (which is what
deletes m_pStatus, dereferencing which the backtrace crashes).

There is code in SwSendMailDialog::dispose() to handle the case that the
xMailDispatcher thread is still running, and stop it without joining,
which would obviously trigger this crash if it were ever executed.

(Maybe i should actually try what happens at runtime, but these
mail-merge dialogs are all scarily confusing.)

Change-Id: I550f6107b064b0c97f3d33bed5bd3830fa2e86f4