sw: speed up the navigator
The call to SwTextBoxHelper::findTextBoxes() in SwDoc::GetFlyNum() made
this so unbelievably slow that in a dbgutil build opening the navigator
on the bugdoc of tdf#94212 the UI freezes because getting all the fly
frames takes longer than the 1 second timeout.

Lets's not retrieve the flys one by one but instead all at once, which
makes it usable again.

Change-Id: Ic41c1648a82dcc3f758ae1b08bac6058f541f25e